tests/requirements.txt
pytest==8.2.0
requests==2.32.3
websocket-client==1.8.0
tenacity==9.0.0
pytest-html==4.1.1
